server/adk/agent.py

from google.adk.agents import LlmAgent

from .planner.agent import planner_agent

# Expose planner as the root agent. The planner contains sub-agents
# for vision, tools analysis, knowledge, and translation. (Commerce removed)
root_agent = planner_agent
